+++ Tests/SpeechTests.cs
@@ -372,5 +372,12 @@ namespace Tests
             SpeechService SpeechService = new SpeechService();
             SpeechService.Speak("Testing drop-off.", null, 50, 1, 30, 40, 0, false);
         }
+
+        [TestMethod]
+        public void TestSectorTranslations()
+        {
+            Assert.AreEqual("Swoiwns N Y dash B a95 dash 0", Translations.StarSystem("Swoiwns NY-B a95-0"));
+            Assert.AreEqual("P P M 5 2 8 7", Translations.StarSystem("PPM 5287"));
+        }
     }
 }